third_party/blink/web_tests/platform/generic/storage/indexeddb/intversion-omit-parameter-expected.txt

Test that initial version after a successful open of a non-existent db is 1

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "intversion-omit-parameter.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
PASS Got upgradeneeded event


openSuccess():
db = event.target.result
Test line from IDBFactory.open: If no version is specified and no database exists, set database version to 1.
PASS db.version is 1
PASS successfullyParsed is true

TEST COMPLETE
